made delete route private through middleware

+++ routes/blogs.js
@@ -5,33 +5,33 @@ var commentRouter = require('./comments');
 var authcontroller = require('../controllers/authcontroller')
 var Comment = require('../models/comment')
 
-router.get('/new', authcontroller.isUserLogged, (req, res, next)=>{
-    res.render('blogform'  )
+router.get('/new', authcontroller.isUserLogged, (req, res, next) => {
+    res.render('blogform')
 })
 
-router.post('/new', authcontroller.isUserLogged, (req, res, next)=>{
+router.post('/new', authcontroller.isUserLogged, (req, res, next) => {
     var tags = req.body.tags.split(',');
     req.body.tags = tags;
-    Blog.create(req.body,(err,result)=>{
-        if(err) return next(err);
+    Blog.create(req.body, (err, result) => {
+        if (err) return next(err);
         res.redirect('/')
     })
 })
 
-router.get('/:id', authcontroller.isUserLogged, (req, res, next)=>{
+router.get('/:id', (req, res, next) => {
     var id = req.params.id;
     Blog.findById(id)
-    .populate({
-        path: 'comments',
-        populate: {
-            path: 'author'
-        }
-    }).populate('author')
-    .exec((err, blog) => {
-        console.log(blog.author, 'check1');
-        if(err) return next(err);
-        res.render('bloglisting', {blog: blog})
-    })
+        .populate({
+            path: 'comments',
+            populate: {
+                path: 'author'
+            }
+        }).populate('author')
+        .exec((err, blog) => {
+            console.log(blog.author, 'check1');
+            if (err) return next(err);
+            res.render('bloglisting', { blog: blog })
+        })
 })
 
 // router.get('/:id', (req,res, next)=>{
@@ -41,37 +41,44 @@ router.get('/:id', authcontroller.isUserLogged, (req, res, next)=>{
 //     })
 // })
 
-router.get('/:id/edit', (req, res, next)=>{
+router.get('/:id/edit', authcontroller.isUserLogged, (req, res, next) => {
     var id = req.params.id;
-    Blog.findById(id, (err, blog)=>{
-        res.render('blogedit', {blog: blog})
+    Blog.findById(id, (err, blog) => {
+        console.log(blog, 'inside blog')
+        res.render('blogedit', { blog: blog })
     })
 })
 
-router.post('/:id/update', (req, res, next)=>{
+router.post('/:id/update', authcontroller.isUserLogged, (req, res, next) => {
     var id = req.params.id;
-    Blog.findByIdAndUpdate(id, req.body, {new:true}, (err,blog)=>{
-        if(err) return next(err);
+    Blog.findByIdAndUpdate(id, req.body, { new: true }, (err, blog) => {
+        if (err) return next(err);
         res.redirect('/')
     })
 })
 
-router.get('/:id/delete', (req,res, next)=>{
+router.get('/:id/delete', authcontroller.isUserLogged, (req, res, next) => {
     var id = req.params.id;
-    Blog.findByIdAndDelete(id, (err, blog)=>{
-        if(err) return next(err);
-        res.redirect('/')
+    Blog.findById(id, (err, blog) => {
+        if(blog.author == req.user._id){
+            Blog.findByIdAndDelete(id, (err,blog)=>{
+                if(err) return next(err);
+                res.redirect('/')
+            })
+        }
+        if (err) return next(err);
+        res.send('not allowed')
     })
 })
 
-router.get('/bloguser', authcontroller.isUserLogged, (req, res, next)=>{
-    if(req.session && req.session.user._id){
-        Blog.find({}, (err, userbloglist)=>{
-            if(err) return next(err);
-            res.render('userblog', {blogitems: userbloglist})
+router.get('/bloguser', authcontroller.isUserLogged, (req, res, next) => {
+    if (req.session && req.session.user._id) {
+        Blog.find({}, (err, userbloglist) => {
+            if (err) return next(err);
+            res.render('userblog', { blogitems: userbloglist })
         })
     }
-    else{
+    else {
         res.redirect('/')
     }
 })
